+++ src/Camera.hpp
@@ -8,8 +8,8 @@
 class Camera
 {
   public:
-  glm::vec3 velocity = glm::vec3(0.0f);
-  glm::vec3 position = glm::vec3(0.0f, 0.3f, 0.0f);
+  glm::dvec3 velocity = glm::vec3(0.0);
+  glm::dvec3 position = glm::vec3(0.0, 0.3, 0.0);
   double pitch { 0.0 };
   double deltaPitch { 0.0 };
   double yaw { 0.0 };
@@ -27,8 +27,8 @@ class Camera
   double oldXpos { 0.0 };
   double oldYpos { 0.0 };
 
-  glm::vec3 forward;
-  glm::vec3 right;
+  glm::dvec3 forward;
+  glm::dvec3 right;
 
   glm::mat4 getViewMatrix();
   glm::mat4 getRotationMatrix();
